feat(valuecard): moved dataState rendering to own component
