+++ src/components/ValueCard/ValueCard.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import withSize from 'react-sizeme';
 import isEmpty from 'lodash/isEmpty';
@@ -7,18 +7,10 @@ import { ValueCardPropTypes, CardPropTypes } from '../../constants/PropTypes';
 import { CARD_LAYOUTS, CARD_SIZES, CARD_CONTENT_PADDING } from '../../constants/LayoutConstants';
 import { COLORS } from '../../styles/styles';
 import Card from '../Card/Card';
+import DataStateRenderer from './DataStateRenderer';
 import { determineMaxValueCardAttributeCount } from '../../utils/cardUtilityFunctions';
 
 import Attribute from './Attribute';
-import { settings } from '../../constants/Settings';
-import classnames from 'classnames';
-import { Tooltip } from 'carbon-components-react';
-import { ErrorFilled24, WarningFilled24 } from '@carbon/icons-react';
-
-const { iotPrefix, prefix: carbonPrefix } = settings;
-const dsPrefix = `${iotPrefix}--data-state`;
-
-export const VALUE_CARD_DATA_STATE = { NO_DATA: 'NO_DATA', ERROR: 'ERROR' };
 
 const ContentWrapper = styled.div`
   display: flex;
@@ -213,92 +205,10 @@ const isLabelAboveValue = (size, layout, attributes, measuredSize) => {
   }
 };
 
-const DataStateIcon = ({ dataState, onClick }) => {
-  let icon;
-  if (dataState.icon) {
-    icon = dataState.icon;
-  } else if (dataState.type === VALUE_CARD_DATA_STATE.ERROR) {
-    icon = <ErrorFilled24 className={`${dsPrefix}-default-error-icon`} />;
-  } else if (dataState.type === VALUE_CARD_DATA_STATE.NO_DATA) {
-    icon = <WarningFilled24 className={`${dsPrefix}-default-warning-icon`} />;
-  }
-
-  const clickWrappedIcon = <span onClick={onClick}>{icon}</span>;
-  return clickWrappedIcon;
-};
-
-const renderDataStateTooltipContent = dataState => {
-  return (
-    <span className={`${dsPrefix}-tooltip`}>
-      <p className={`${dsPrefix}-tooltip__label`}>{dataState.label}</p>
-      <p>{dataState.description}</p>
-      <p>{dataState.extraTooltipText}</p>
-      {dataState.learnMoreURL ? (
-        <a
-          className={`${carbonPrefix}--link`}
-          href={dataState.learnMoreURL}
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          {dataState.learnMoreText}
-        </a>
-      ) : null}
-    </span>
-  );
-};
-
 const ValueCard = ({ title, content, size, values, isEditable, i18n, dataState, ...others }) => {
-  const [tooltipOpen, setTooltipOpen] = useState(false);
-  const toggleTooltip = () => setTooltipOpen(open => !open);
   const availableActions = {
     expand: false,
     ...others.availableActions,
-  };
-
-  const renderElementWithTooltip = triggerElement => {
-    return (
-      <Tooltip open={tooltipOpen} showIcon={false} triggerText={triggerElement}>
-        {renderDataStateTooltipContent(dataState)}
-      </Tooltip>
-    );
-  };
-
-  const renderDataStateGridItems = () => {
-    const label = (
-      <span className={`${dsPrefix}-grid__label`} onClick={toggleTooltip}>
-        {dataState.label}
-      </span>
-    );
-    return (
-      <React.Fragment>
-        <DataStateIcon dataState={dataState} onClick={toggleTooltip} />
-        {renderElementWithTooltip(label)}
-        <p className={`${dsPrefix}-grid__description`} onClick={toggleTooltip}>
-          {dataState.description}
-        </p>
-      </React.Fragment>
-    );
-  };
-
-  const renderDataStateGridItemsXSmall = () => {
-    const icon = <DataStateIcon dataState={dataState} onClick={toggleTooltip} />;
-    return renderElementWithTooltip(icon);
-  };
-
-  const renderDataState = () => {
-    return (
-      <div
-        className={`${dsPrefix}-container`}
-        style={{ '--container-padding': `${CARD_CONTENT_PADDING}px` }}
-      >
-        <p className={classnames(`${dsPrefix}-dashes`)}>--</p>
-        <div className={`${dsPrefix}-grid`}>
-          {size === CARD_SIZES.XSMALL
-            ? renderDataStateGridItemsXSmall()
-            : renderDataStateGridItems()}
-        </div>
-      </div>
-    );
   };
 
   return (
@@ -329,7 +239,7 @@ const ValueCard = ({ title, content, size, values, isEditable, i18n, dataState, 
             {...others}
           >
             <ContentWrapper layout={layout}>
-              {isEmpty(values) && dataState && renderDataState(dataState)}
+              {dataState && <DataStateRenderer dataState={dataState} size={size} />}
               {!dataState &&
                 attributes.map((attribute, i) => (
                   <React.Fragment key={`fragment-${attribute.dataSourceId}`}>
