+++ src/constants/PropTypes.js
@@ -2,8 +2,15 @@ import PropTypes from 'prop-types';
 
 import { bundledIconNames } from '../utils/bundledIcons';
 
-import { CARD_SIZES, CARD_LAYOUTS, DASHBOARD_SIZES, TIME_SERIES_TYPES } from './LayoutConstants';
-import { VALUE_CARD_DATA_STATE } from '../components/ValueCard/ValueCard';
+import {
+  CARD_SIZES,
+  CARD_LAYOUTS,
+  DASHBOARD_SIZES,
+  TIME_SERIES_TYPES,
+  VALUE_CARD_DATA_STATE,
+} from './LayoutConstants';
+// import { VALUE_CARD_DATA_STATE } from '../components/ValueCard/ValueCard';
+// import { VALUE_CARD_DATA_STATE } from '../components/ValueCard/ValueCard';
 
 export const AttributePropTypes = PropTypes.shape({
   label: PropTypes.string, // optional for little cards
